requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-paypal==0.4.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.6
idna==2.6
Pillow==5.0.0
pytz==2017.3
requests==2.18.4
six==1.11.0
urllib3==1.22
